mcnode refactor: split out protocol specific code
